+++ src/app/components/einsaetze/zeitachse/service/timeaxisCalculator/timeaxisCalculator.js
